+++ src/AdminResources/Files.php
@@ -28,6 +28,7 @@ class Files extends MsGraphAdmin
 
     public function setUser($userId = null)
     {
+<<<<<<< HEAD
         $userDriveId =  MsConfig::get('drive_account');
         if(!$userDriveId && !$userId) {
             throw new \ApplicationException('Pour un utilisateur il faut soit le compte principal dans le drive, soit passer l\'id dans les paramètres de userId()');
@@ -37,13 +38,17 @@ class Files extends MsGraphAdmin
             $this->userId = UserModel::find($userDriveId)->msgraph_id;
         }
         $this->baseRequest = 'users/'.$this->userId.'/';
+=======
+        $this->userId = $userId;
+
+>>>>>>> master
         return $this;
     }
 
     public function getDrives()
     {
         if ($this->userId == null) {
-            throw new Exception("userId is required.");
+            throw new Exception('userId is required.');
         }
 
         return MsGraphAdmin::get($this->baseRequest.'/drives');
@@ -70,9 +75,15 @@ class Files extends MsGraphAdmin
 
     public function downloadFile($id)
     {
+<<<<<<< HEAD
         // if ($this->userId == null) {
         //     throw new Exception("userId is required.");
         // }
+=======
+        if ($this->userId == null) {
+            throw new Exception('userId is required.');
+        }
+>>>>>>> master
 
         $id = MsGraphAdmin::get($this->baseRequest.'/drive/items/'.$id);
 
@@ -149,6 +160,7 @@ class Files extends MsGraphAdmin
 
     public function uploadBigFile($name, $uploadPath, $path = null)
     {
+<<<<<<< HEAD
         $finalPath = $path ? $this->baseFolder.'/'.$path : $this->baseFolder;
         $uploadSession = $this->createUploadSession($name, $finalPath);
         $uploadUrl = $uploadSession['uploadUrl'];
@@ -213,6 +225,10 @@ class Files extends MsGraphAdmin
     {
         if (substr($string, 0, 1) !== "/") {
             $string = "/$string";
+=======
+        if ($this->userId == null) {
+            throw new Exception('userId is required.');
+>>>>>>> master
         }
 
         return $string;