+++ src/Models/MsGraphToken.php
@@ -7,4 +7,9 @@ use Illuminate\Database\Eloquent\Model;
 class MsGraphToken extends Model
 {
     protected $guarded = [];
+
+    public function __construct()
+    {
+        $this->setConnection(config('msgraph.dbConnection'));
+    }
 }