+++ changelog.md
@@ -271,3 +271,16 @@ The in a view the previous and next links can be displayed:
 	<a href='{{ url()->current().'?next='.$next }}'>Next Page</a>
 @endif
 ```
+
+### 3.1.2
+
+added support for Laravel 9
+
+### 3.1.3
+
+Added files import for MsGraphAdmin 
+
+### 3.1.4
+
+Added test foundation
+Added php code sniffer fixer and style config